% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearFit.R
\name{linearFit}
\alias{linearFit}
\title{linearFit}
\usage{
linearFit(data, col_dil = "dilution_conc", col_res = "dilution_resp",
  col_uid = "uid", rjct_pre = 0.25, rjct_acc = 0.25, rjct_rsq = 0.9,
  rjct_pts = 0.75, filter = T)
}
\arguments{
\item{data}{data.frame of dilution~response values}

\item{col_dil}{column containing the variable for dilution concentration}

\item{col_res}{column containing the variable for dilution response}

\item{col_uid}{column containing the variable for unique id}

\item{rjct_pre}{rejection criteria for min precision, default 0.25,}

\item{rjct_acc}{rejection criteria for min accuracy, default 0.25,}

\item{rjct_rsq}{rejection criteria for min linearity Pearson corrleation, default 0.90,}

\item{rjct_pts}{rejection criteria for dilution points retained, default 0.75,}

\item{filter}{filter dataset to top solution passing reject criteria, default TRUE}
}
\description{
linearFit
}
\examples{
linearFit
}
\keyword{linearFit}
