\name{OH}
\alias{OH}
\docType{data}
\title{
OH dataset
}
\description{
The hydroxyl \code{OH} group activity of compounds from molecular descriptors.
}
\usage{data(OH)}
\format{
  A data frame with 719 observations on 294 descriptors/predictors. 
  The response is \code{act}.
}

\source{
The dataset was provided by Tomas Oberg.
}

\examples{
data(OH)
}
\keyword{datasets}
