\name{orthonorm}
\alias{orthonorm}
\title{
Gram-Schmidt orthonormalization
}
\description{
Function to obtain the Gram-Schmidt orthonormalization of a matrix.
}
\usage{
orthonorm(u)
}
\arguments{
  \item{u}{
A matrix.
}
}

\references{
James R. Schott (1997). Matrix Analysis for Statistics. Wiley Series in Probability and Statistics
}
\author{
Kofi Placid Adragni
}
\keyword{internal}

