\name{plot.pfc}
\alias{plot.pfc}
\title{
Internal function of \code{ldr}
}
\description{
A plot method for an object of class \code{pfc}.
}
\usage{
\method{plot}{pfc}(x, ...)
}
\arguments{
  \item{x}{
An object of class \code{pfc}.
}
  \item{\dots}{
Other arguments for plot method.
}
}

\author{
Kofi Placid Adragni <kofi@umbc.edu>
}
\keyword{internal} 