\name{print.lad}
\alias{print.lad}
\title{
Internal function of \code{ldr}
}
\description{
A plot method for an object of class \code{lad}.
}
\usage{
\method{print}{lad}(x,...)
}
\arguments{
  \item{x}{
An object of class \code{lad}.
}
  \item{\dots}{
Other arguments for plot method.
}
}
\author{
Kofi Placid Adragni <kofi@umbc.edu>
}
\keyword{internal} 