\name{print.ldr}
\alias{print}
\alias{print.ldr}
\title{
Internal function for ldr.
}
\description{
Internal function to print ldr object.
}
\usage{
\method{print}{ldr}(x,...)
}
\arguments{
  \item{x}{
An object of class \code{ldr}, fit with \code{pfc}, \code{lad}, or \code{core}.
}
  \item{...}{
Other arguments to generic print method.
}
}
\keyword{internal}
