\name{print.pfc}
\alias{print.pfc}
\title{
Internal function of \code{ldr}
}
\description{
A print method for an object of class \code{pfc}.
}
\usage{
\method{print}{pfc}(x,...)
}
\arguments{
  \item{x}{
An object of class \code{pfc}.
}
  \item{\dots}{
Other arguments for plot method.
}
}
\author{
Kofi Placid Adragni <kofi@umbc.edu>
}
\keyword{internal} 