\name{print.structure}
\alias{print.structure}
\title{
Internal function for ldr.
}
\description{
Internal function to print a the test of structure of \code{pfc} objects.
}
\usage{
\method{print}{structure}(x,...)
}
\arguments{
  \item{x}{
An object of class \code{structure} obtained using the function \code{structure.test}.
}
}
\author{
Kofi Placid Adragni <kofi@umbc.edu>
}
\keyword{internal} 