\name{print.summary.ldr}
\alias{print.summary.ldr}
\title{
Internal function for ldr.
}
\description{
Internal function to print an ldr summary object.
}
\usage{
\method{print}{summary.ldr}(x,...)
}
\arguments{
  \item{object}{
A summary object of class \code{sldr}.
}
  \item{...}{
Other arguments to generic summary method.
}
}
\keyword{internal}

