\name{print.summarycore}
\alias{print.summarycore}
\title{
Internal function for ldr.
}
\description{
Internal function to print a \code{core} object.
}
\usage{
\method{print}{summarycore}(x,...)
}
\arguments{
  \item{x}{
An object of class \code{summarycore}.
}
}
\author{
Kofi Placid Adragni <kofi@umbc.edu>
}
\keyword{internal} 