\name{print.summarylad}
\alias{print.summarylad}
\title{
Internal function for ldr.
}
\description{
Internal function to print a \code{lad} object.
}
\usage{
\method{print}{summarylad}(x,...)
}
\arguments{
  \item{x}{
An object of class \code{lad}.
}
}
\author{
Kofi P. Adragni <kofi@umbc.edu>
}
\keyword{internal}


