\name{print.summarypfc}
\alias{print.summarypfc}

\title{
Internal function for ldr.
}
\description{
Internal function to print a \code{pfc} object.
}
\usage{
\method{print}{summarypfc}(x,...)
}
\arguments{
  \item{x}{
An object of class \code{summarypfc}.
}
}
\author{
Kofi Placid Adragni <kofi@umbc.edu>
}
\keyword{internal} 