\name{summary.core}
\alias{summary.core}
\title{
Internal function for ldr.
}
\description{
Internal function for the summary output of a \code{core} object.
}
\usage{
\method{summary}{core}(object,...)
}
\arguments{
  \item{object}{
An object of class \code{core}.
}
}
\author{
Kofi Placid Adragni <kofi@umbc.edu>
}
\keyword{internal}
