\name{summary.lad}
\alias{summary.lad}
\title{
Internal function for ldr.
}
\description{
Internal function for the summary output of a \code{lad} object.
}
\usage{
\method{summary}{lad}(object,...)
}
\arguments{
  \item{object}{
An object of class \code{lad}.
}
}
\author{
Kofi Placid Adragni <kofi@umbc.edu>
}
\keyword{internal} 