\name{summary.ldr}
\alias{summary}
\alias{summary.ldr}
\title{
Summary method for object of class \code{ldr}
}
\description{
A summary method for object of class \code{ldr}, including \code{pfc}, \code{lad}, and \code{core}.
}
\usage{
\method{summary}{ldr}(object, y=NULL,...)
}
\arguments{
  \item{object}{
An object of class \code{ldr}, fit with \code{pfc}, \code{lad}, or \code{core}.
}
  \item{y}{
The response vector for additional \code{pfc} log.
}
  \item{...}{
Other arguments to generic summary method.
}
}
\author{
Kofi Placid Adragni <kofi@umbc.edu>
}

