\name{summary.pfc}
\alias{summary.pfc}
\title{
Internal function for ldr.
}
\description{
Internal function for the summary output of a \code{pfc} object.
}
\usage{
\method{summary}{pfc}(object,...)
}
\arguments{
  \item{object}{An object of class \code{pfc}.}
}
\author{
Kofi Placid Adragni <kofi@umbc.edu>
}
\keyword{internal} 