\name{print.structure}
\alias{structure}
\alias{print.structure}
\title{
Internal function for ldr.
}
\description{
Internal function to print structure object of pfc.
}
\usage{
\method{print}{structure}(x,...)
}
\arguments{
  \item{x}{
An object of class \code{ldr}, fit with \code{pfc}, \code{lad}, or \code{core}.
}
  \item{...}{
Other arguments to generic print method.
}
}
\keyword{internal}
